make a separate card component
PlayersGrid.tsx became too big with card, add, and remove lifepoint
buttons.
